Avoid copying strings in attributedTextWithLinks.
We don't need to make a local mutable (and autoreleased) copy of our
attributed string in attributedTextWithLinks. Instead, we can operate
directly on our _attributedText ivar.

This change also adds some early-out conditions that let us avoid the
bulk of the processing code when we don't have any links to set.